rstracer: remove tech table from vacuum

+++ rstracer/src/pipeline/stage/vacuum.rs
@@ -6,7 +6,11 @@ pub fn request(config: VacuumConfig, schema: Schema) -> String {
 
     for table in schema.tables {
         for layer in config.to_list() {
-            if table.name.starts_with(&layer.0) && layer.1 > 0 && !table.name.contains("_dim_") {
+            if table.name.starts_with(&layer.0)
+                && layer.1 > 0
+                && !table.name.contains("_dim_")
+                && !table.name.contains("_tech_")
+            {
                 query.push_str(&format!(
                     "BEGIN; DELETE FROM memory.{} WHERE inserted_at + '{} seconds' < CURRENT_TIMESTAMP; COMMIT;",
                     table.name, layer.1
